Clean up sample postgres settings to work better out of the box.
The Mac homebrew install of PostgreSQL and the Ubuntu package
use different locations for the unix domain socket. If we specify
the socket directory as /tmp, that works on Mac but not Ubuntu,
which expects /var/run/postgresql.

By omitting the path entirely, the postgres client will automatically
choose the correct socket directory path for each environment. Win!